@@ -26,10 +26,10 @@
     "ietf-language-tag-regex": "^0.0.5",
     "inversify": "^5.0.1",
     "js-yaml": "^3.13.1",
-    "log4js": "^6.1.1",
+    "log4js": "^6.1.2",
     "moment": "^2.24.0",
     "mysql2": "^2.1.0",
-    "pg": "^7.18.1",
+    "pg": "^7.18.2",
     "plural-rules": "^1.0.1",
     "reflect-metadata": "^0.1.13",
     "socks-proxy-agent": "^5.0.0",
@@ -41,23 +41,23 @@
   "devDependencies": {
     "@types/cron": "^1.7.1",
     "@types/eth-sig-util": "^2.1.0",
-    "@types/jest": "^25.1.1",
+    "@types/jest": "^25.1.3",
     "@types/js-yaml": "^3.12.2",
-    "@types/node": "^13.7.0",
+    "@types/node": "^13.7.4",
     "@types/table": "^4.0.7",
-    "@typescript-eslint/eslint-plugin": "^2.19.0",
-    "@typescript-eslint/parser": "^2.19.0",
+    "@typescript-eslint/eslint-plugin": "^2.20.0",
+    "@typescript-eslint/parser": "^2.20.0",
     "eslint": "^6.8.0",
-    "firebase-tools": "^7.12.1",
-    "husky": "^4.2.1",
+    "firebase-tools": "^7.13.1",
+    "husky": "^4.2.3",
     "jest": "^25.1.0",
     "sqlite3": "^4.1.1",
-    "ts-jest": "^25.2.0",
+    "ts-jest": "^25.2.1",
     "ts-node": "^8.6.2",
     "ts-node-dev": "^1.0.0-pre.44",
     "tsconfig-paths": "^3.9.0",
-    "typescript": "^3.7.5",
-    "vuepress": "^1.3.0",
+    "typescript": "^3.8.2",
+    "vuepress": "^1.3.1",
     "vuepress-plugin-autobar": "boboidream/vuepress-bar"
   }
 }
